api/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
firebase
python-dotenv
SpeechRecognition
python-jwt
gcloud
sseclient
Crypto
pycryptodome
requests_toolbelt
tensorflow
librosa
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
firebase==3.0.1
firebase-admin==4.3.0
Flask==1.1.2
google-api-core==1.20.0
google-api-python-client==1.9.2
google-auth==1.16.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.29.0
httplib2==0.18.1
idna==2.9
msgpack==1.0.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.13.0
python-firebase==1.2
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.9
